requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.1
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
openai==1.3.0
langdetect==1.0.9
googletrans==4.0.0-rc1
reportlab==4.0.4
python-docx==0.8.11
jwt==1.3.1
PyJWT==2.7.0
